make auth.publicurl required and add stricter validation
