+++ core/url.go
@@ -18,7 +18,13 @@
 
 package core
 
-import "strings"
+import (
+	"errors"
+	"net"
+	"net/url"
+	"slices"
+	"strings"
+)
 
 // JoinURLPaths works like path.Join but for URLs; it won't remove double slashes.
 // It makes sures there is only one slash between the parts.
@@ -34,4 +40,60 @@ func JoinURLPaths(parts ...string) string {
 		result = strings.TrimSuffix(result, "/") + "/" + strings.TrimPrefix(parts[i], "/")
 	}
 	return result
+}
+
+// ParsePublicURL parses the given input string as URL and asserts that
+// it has a scheme,
+// it is not an IP address, and
+// it is not a reserved address or TLD as described in RFC2606 or https://www.ietf.org/archive/id/draft-chapin-rfc2606bis-00.html.
+func ParsePublicURL(input string) (*url.URL, error) {
+	if !strings.Contains(input, "://") {
+		return nil, errors.New("URL missing scheme")
+	}
+	parsed, err := url.Parse(input)
+	if err != nil {
+		return nil, err
+	}
+	if net.ParseIP(parsed.Hostname()) != nil {
+		return nil, errors.New("hostname is IP")
+	}
+	if isReserved(parsed) {
+		return nil, errors.New("hostname is reserved")
+	}
+	return parsed, nil
+}
+
+// isReserved returns true if URL uses any of the reserved TLDs or addresses
+func isReserved(URL *url.URL) bool {
+	parts := strings.Split(strings.ToLower(URL.Hostname()), ".")
+	tld := parts[len(parts)-1]
+	if slices.Contains(reservedTLDs, tld) {
+		return true
+	}
+
+	if len(parts) > 1 {
+		l2address := strings.Join(parts[len(parts)-2:], ".")
+		return slices.Contains(reservedAddresses, l2address)
+	}
+
+	return false
+}
+
+var reservedTLDs = []string{
+	"", // no domain specified
+	"corp",
+	"example",
+	"home",
+	"host",
+	"invalid",
+	"lan",
+	"local",
+	"localdomain",
+	"localhost",
+	"test",
+}
+var reservedAddresses = []string{
+	"example.com",
+	"example.net",
+	"example.org",
 }
