+++ network/transport/types_test.go
@@ -19,53 +19,12 @@
 package transport
 
 import (
-	"errors"
 	"github.com/nuts-foundation/nuts-node/core"
 	"testing"
 
 	"github.com/nuts-foundation/go-did/did"
 	"github.com/stretchr/testify/assert"
 )
-
-func Test_ParseAddress(t *testing.T) {
-	errScheme := errors.New("scheme must be grpc")
-	errIsIp := errors.New("hostname is IP")
-	errIsReserved := errors.New("hostname is reserved")
-
-	tests := []struct {
-		input  string
-		output string
-		err    error
-	}{
-		{"grpc://foo.bar:5050", "foo.bar:5050", nil},
-		{"grpc://foo.BAR", "foo.BAR", nil},
-		{"grpc://fooBAR", "fooBAR", nil},
-		{"foo.bar", "", errScheme},
-		{"http://foo.bar", "", errScheme},
-		{"grpc://1.2.3.4:5555", "", errIsIp},
-		{"grpc://[::1]:5555", "", errIsIp},
-		{"grpc://localhost", "", errIsReserved},
-		{"grpc://my.localhost:5555", "", errIsReserved},
-		{"grpc://my.local", "", errIsReserved},
-		{"grpc://LOCALhost:5555", "", errIsReserved},
-		{"grpc://:5555", "", errIsReserved},
-		{"grpc://example.com", "", errIsReserved},
-		{"grpc://my.example.com:5555", "", errIsReserved},
-	}
-
-	for _, tc := range tests {
-		addr, err := parseNutsCommAddress(tc.input)
-		if tc.err == nil {
-			// valid test cases
-			assert.Equal(t, tc.output, addr.Host, "test case: %v", tc)
-			assert.NoError(t, err, "test case: %v", tc)
-		} else {
-			// invalid test cases
-			assert.Empty(t, addr, "test case: %v", tc)
-			assert.EqualError(t, err, tc.err.Error(), "test case: %v", tc)
-		}
-	}
-}
 
 func TestPeer_ToFields(t *testing.T) {
 	peerLogFields := Peer{
@@ -92,7 +51,7 @@ func TestNutsCommURL_UnmarshalJSON(t *testing.T) {
 
 	t.Run("error - invalid url, missing grpc", func(t *testing.T) {
 		var url NutsCommURL
-		err := url.UnmarshalJSON([]byte(`"foo.bar:5050"`))
+		err := url.UnmarshalJSON([]byte(`"https://foo.bar:5050"`))
 		assert.EqualError(t, err, "scheme must be grpc")
 	})
 
@@ -100,6 +59,5 @@ func TestNutsCommURL_UnmarshalJSON(t *testing.T) {
 		var url NutsCommURL
 		err := url.UnmarshalJSON([]byte(`123`))
 		assert.EqualError(t, err, "endpoint not a string")
-
 	})
 }
